cleanup of unused symbols and final refactoring, added more comments, moved data model to separate header file
