+++ src/graphdatamodel.h
@@ -1,0 +1,14 @@
+/* GTlab - Gas Turbine laboratory
+ *
+ * SPDX-License-Identifier: MPL-2.0+
+ * SPDX-FileCopyrightText: 2024 German Aerospace Center (DLR)
+ *
+ * Created on: 8.10.2024
+ * Author: Marius Bröcker (AT-TWK)
+ * E-Mail: marius.broecker@dlr.de
+ */
+
+#ifndef GRAPHDATAMODEL_H
+#define GRAPHDATAMODEL_H
+
+#endif // GRAPHDATAMODEL_H
